Add solution for problem `B - Alyona and Mex` Link: https://codeforces.com/contest/682/problem/B Tags: sortings, *1200 Ref: https://codeforces.com/contest/682/submission/74647719

+++ codeforces/682/B.cpp
@@ -3,7 +3,7 @@ using namespace std;
 int main(){
   int n;
   cin>>n;
-  int a[n+1],ans=n+1;
+  long long int a[n+1],ans=n+1;
   for(int i=0;i<n;i++)
     cin>>a[i];
   sort(a,a+n);
